Fixed the way enemies spawn
Now enemies will spawn on the outside of the player along a path.

Also, added a function in game.gd to adjust the rate of enemy spawns. Now under the process function in game.gd you can control a timer that changes what enemies spawn at different times of the game. Currently only goes up to 100 seconds in terms of changing spawns.