added green chili 😃
green chili is like the red chili, except it fires a stream of 6 bullets really fast that do less damage, and then has a short cool down before firing again.

i made duplicates of all the chili stuff and grabbed a simple shader to change the color of the bullets.